use `git merge-base <base_branch> HEAD` to identify the `from` commit
Do not rely on an HTTP API to identify the from commit as the CI build
is launched before the PR is created. So we can't know what are the
commits of the PR.

Instead, we rely on `git merge-base` to find the best ancestor for a
merge (i.e. the `from` commit of the range of commits of the PR).

Also use `origin/master` as the default base-branch param to use the
remote repo (source of truth).