split app into different packages to speed up compilation
Split the app into different packages. Prior to this change, the app was
contained in a single package, which was related to all the
dependencies (e.g. `structopt`). Each single change on the codebase will
trigger a full recompilation of the single package, which was long (> 15s),
and so a pain to develop with.

Now that the packages are split, the code changes only trigger a
recompilation of the changed package, which speeds up the whole
compilation process (compilation on changes only + linking). The compile
time is now < 1s.

Related article : https://medium.com/@jondot/my-key-learnings-after-30-000-loc-in-rust-a553e6403c19