differenciate remote and base branch parameters
Differenciate the remote (defaults to `origin`) and base_branch
(defaults to `master`) parameters in order to identify whether we're on
the `origin/master` branch (i.e. when a merge is done on master) or when
we're on `origin/feature-branch` (i.e. when a commit is pushed on a
feature branch).

Mixing the remote and the branch in the `base-branch` parameter
(e.g. `origin/master`) wasn't precise enough to check when we've
checked out the default base branch from the default remote, as we were
only considering the current branch name and not the current remote name
during comparison.