fix circleci cache archive
There is a validation error on the circleci cache archive :
https://circleci.com/gh/KnpLabs/should-skip-ci/16

The cache archives with the Cargo.lock file before and after these change
are invalid to circleci. We bump the cache archive version to create a
fresh one.

The build on this PR should be ran twice : once to create and upload the cache
archive, and once to use it.