Add a functional test covering a master and develop branches based git workflow
This test file is meant to test a `master` and `develop` branches git based
workflow.
In this workflow, the `master` branch is ISO to the prod, and the `develop`
branch contains the latest features.
Developers usually issue new feature branches from the `develop` branch.
The workflow is represented by the following schema :

```
master ---o-------o
           \     /
develop     o---o
             \ /
feature1      o

---> direction
```