fix automatic assets upload when a release is published
The previous token was out of date. I created a new machine user on
github with access to this repo only, and generated a new access token.
I updated circleci env vars with this token.
I manually ran the `upload_release_asset.sh` script on circleci to test
these settings and to upload the assets of v0.2.1 release (which has
previously failed due to previous credentials invalidity).