fix ssc default behavior : must use base branch comparison when no optional args are provided
PR [40](https://github.com/KnpLabs/should-skip-ci/pull/40) has
introduced a change in the default behavior of ssc. When running the
tool as usual, there was an error raised :

```
$ ssc --path <path> --cmd <stop_command>

error: a value is required for '--base-ref <BASE_REF>' but none was supplied
```

whereas it should use the base branch comparison by default (as we don't
want people who are upgrading ssc to change their CI scripts).

This commit fixes this issue.